+++ src/lib/models/show/show-simplified.ts
